requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.3.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
numpy==1.23.5
pandas==1.5.3
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
typing_extensions==4.5.0
Werkzeug==2.3.3
zipp==3.15.0